split pkg/cmd
