+++ pkg/cmd/root_test.go
@@ -211,16 +211,6 @@ var testNodeMetrics = &metricsapiv1beta1.NodeMetricsList{
 	},
 }
 
-// var testNodeMetrics = &metricsapiv1beta1.NodeMetrics{
-// 	ObjectMeta: metav1.ObjectMeta{
-// 		Name: "node1",
-// 	},
-// 	Usage: v1.ResourceList{
-// 		v1.ResourceCPU:    *resource.NewMilliQuantity(100, resource.DecimalSI),
-// 		v1.ResourceMemory: *resource.NewQuantity(1024, resource.DecimalSI),
-// 	},
-// }
-
 func TestNewFreeOptions(t *testing.T) {
 	streams := genericclioptions.IOStreams{In: os.Stdin, Out: os.Stdout, ErrOut: os.Stderr}
 
@@ -726,235 +716,6 @@ func TestPrepareListTableHeader(t *testing.T) {
 		})
 	}
 
-}
-
-func TestShowFree(t *testing.T) {
-
-	var tests = []struct {
-		description string
-		pod         bool
-		namespace   string
-		noheader    bool
-		nometrics   bool
-		expected    []string
-		expectedErr error
-	}{
-		{
-			"default free",
-			false,
-			"default",
-			true,
-			true,
-			[]string{
-				"node1   Ready   1     2     4     25%   50%   1K    2K    4K    25%   50%",
-				"",
-			},
-			nil,
-		},
-		{
-			"default free with metrics",
-			false,
-			"default",
-			true,
-			false,
-			[]string{
-				"node1   Ready   -     1     2     4     0%    25%   50%   -     1K    2K    4K    0%    25%   50%",
-				"",
-			},
-			nil,
-		},
-		{
-			"default free --pod",
-			true,
-			"default",
-			true,
-			true,
-			[]string{
-				"node1   Ready   1     2     4     25%   50%   1K    2K    4K    25%   50%   1     110   1",
-				"",
-			},
-			nil,
-		},
-		{
-			"awesome-ns free",
-			true,
-			"awesome-ns",
-			true,
-			true,
-			[]string{
-				"node1   Ready   200m   200m   4     5%    5%    0K    0K    4K    7%    7%    1     110   2",
-				"",
-			},
-			nil,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.description, func(t *testing.T) {
-
-			fakeNodeClient := fake.NewSimpleClientset(&testNodes[0])
-			fakePodClient := fake.NewSimpleClientset(&testPods[0], &testPods[2])
-			fakeMetricsNodeClient := fakemetrics.NewSimpleClientset(&testNodeMetrics.Items[0])
-			fakeMetricsPodClient := fakemetrics.NewSimpleClientset(testPodMetrics)
-
-			buffer := &bytes.Buffer{}
-			o := &FreeOptions{
-				nocolor:           true,
-				table:             table.NewOutputTable(buffer),
-				list:              false,
-				pod:               test.pod,
-				noHeaders:         true,
-				noMetrics:         test.nometrics,
-				nodeClient:        fakeNodeClient.CoreV1().Nodes(),
-				podClient:         fakePodClient.CoreV1().Pods(test.namespace),
-				metricsPodClient:  fakeMetricsPodClient.MetricsV1beta1().PodMetricses("default"),
-				metricsNodeClient: fakeMetricsNodeClient.MetricsV1beta1().NodeMetricses(),
-			}
-
-			if err := o.showFree([]v1.Node{testNodes[0]}); err != nil {
-				t.Errorf("unexpected error: %v", err)
-				return
-			}
-
-			e := strings.Join(test.expected, "\n")
-			if buffer.String() != e {
-				t.Errorf("expected(%s) differ (got: %s)", e, buffer.String())
-				return
-			}
-
-		})
-	}
-
-	t.Run("Allnamespace", func(t *testing.T) {
-
-		fakeNodeClient := fake.NewSimpleClientset(&testNodes[0])
-		fakePodClient := fake.NewSimpleClientset(&testPods[0], &testPods[1], &testPods[2])
-		fakeMetricsNodeClient := fakemetrics.NewSimpleClientset(&testNodeMetrics.Items[0])
-		fakeMetricsPodClient := fakemetrics.NewSimpleClientset(testPodMetrics)
-
-		buffer := &bytes.Buffer{}
-		o := &FreeOptions{
-			nocolor:           true,
-			table:             table.NewOutputTable(buffer),
-			list:              false,
-			pod:               false,
-			allNamespaces:     true,
-			noHeaders:         true,
-			noMetrics:         true,
-			nodeClient:        fakeNodeClient.CoreV1().Nodes(),
-			podClient:         fakePodClient.CoreV1().Pods(""),
-			metricsPodClient:  fakeMetricsPodClient.MetricsV1beta1().PodMetricses("default"),
-			metricsNodeClient: fakeMetricsNodeClient.MetricsV1beta1().NodeMetricses(),
-		}
-
-		if err := o.showFree([]v1.Node{testNodes[0]}); err != nil {
-			t.Errorf("unexpected error: %v", err)
-			return
-		}
-
-		expected := []string{
-			"node1   Ready   1700m   2700m   4     42%   67%   2K    3K    4K    57%   82%",
-			"",
-		}
-		e := strings.Join(expected, "\n")
-		if buffer.String() != e {
-			t.Errorf("expected(%s) differ (got: %s)", e, buffer.String())
-			return
-		}
-
-	})
-}
-
-func TestListPodsOnNode(t *testing.T) {
-
-	var tests = []struct {
-		description   string
-		listContainer bool
-		listAll       bool
-		nometrics     bool
-		expected      []string
-	}{
-		{
-			"list container with metrics",
-			false,
-			false,
-			false,
-			[]string{
-				"node2   default   pod2   <unknown>   2.3.4.5   Running   container2a   -     500m   500m   -     1K    1K",
-				"",
-			},
-		},
-		{
-			"list container: false, list all: false",
-			false,
-			false,
-			true,
-			[]string{
-				"node2   default   pod2   <unknown>   2.3.4.5   Running   container2a   500m   500m   1K    1K",
-				"",
-			},
-		},
-		{
-			"list container: true, list all: false",
-			true,
-			false,
-			true,
-			[]string{
-				"node2   default   pod2   <unknown>   2.3.4.5   Running   container2a   500m   500m   1K    1K    nginx:latest",
-				"",
-			},
-		},
-		{
-			"list container: false, list all: true",
-			false,
-			true,
-			true,
-			[]string{
-				"node2   default   pod2   <unknown>   2.3.4.5   Running   container2a   500m   500m   1K    1K",
-				"node2   default   pod2   <unknown>   2.3.4.5   Running   container2b   -      -      -     -",
-				"",
-			},
-		},
-		{
-			"list container: true, list all: true",
-			true,
-			true,
-			true,
-			[]string{
-				"node2   default   pod2   <unknown>   2.3.4.5   Running   container2a   500m   500m   1K    1K    nginx:latest",
-				"node2   default   pod2   <unknown>   2.3.4.5   Running   container2b   -      -      -     -     busybox:latest",
-				"",
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.description, func(t *testing.T) {
-			buffer := &bytes.Buffer{}
-			fakeClient := fake.NewSimpleClientset(&testPods[1])
-			o := &FreeOptions{
-				table:              table.NewOutputTable(buffer),
-				podClient:          fakeClient.CoreV1().Pods(""),
-				noHeaders:          true,
-				noMetrics:          test.nometrics,
-				nocolor:            true,
-				listContainerImage: test.listContainer,
-				listAll:            test.listAll,
-			}
-
-			if err := o.listPodsOnNode([]v1.Node{testNodes[1]}); err != nil {
-				t.Errorf("unexpected error: %v", err)
-				return
-			}
-
-			expected := strings.Join(test.expected, "\n")
-			actual := buffer.String()
-			if actual != expected {
-				t.Errorf("[%s] expected(%s) differ (got: %s)", test.description, expected, actual)
-				return
-			}
-		})
-	}
 }
 
 func TestToUnit(t *testing.T) {
